chore(ci.yml): update runs-on to use ubuntu-latest and arangodb image to 3.10.9
The runs-on field in the ci.yml file has been updated to use
ubuntu-latest instead of ubuntu-20.04. The arangodb image has been
updated to version 3.10.9. The go-version field has been updated to use
~1.20.6. The actions/cache version has been updated to v3. The docker
build and push tags have been updated to include the branch name in
addition to the short sha.