chore(testcov.yml): update arangodb image to version 3.10.9 for compatibility and bug fixes
The runs-on field is updated to use ubuntu-latest instead of a specific
version of Ubuntu to ensure better compatibility with different versions
of GitHub Actions runners. The arangodb image is updated to version
3.10.9 to take advantage of compatibility improvements and bug fixes.
The go-version field is updated to use a range (~1.20.6) instead of a
specific version to allow for better compatibility with different
versions of Go.