ci(ci.yml): update Docker workflow to use docker-publish and specify image and Dockerfile
The GitHub Actions CI configuration now uses the `docker-publish.yaml`
workflow instead of `docker-build-push.yaml` to align with updated CI
practices. Additionally, specifying the `image` and `Dockerfile`
explicitly in the workflow ensures that the correct Docker image is
built and published, improving the automation and clarity of the CI/CD
process.