feat(collection): add generic Filter function to filter slices
This commit introduces a generic `Filter` function to the `collection`
package. This function allows filtering a slice of any type `T` based on
a provided predicate function. The predicate function takes an element
of type `T` and returns a boolean indicating whether the element should
be included in the resulting filtered slice.