feat(collection): add TFold and CurriedTFold functions
The TFold and CurriedTFold functions are added to provide a way to fold
over a Tuple2, which is a pair of values with independent types. This
allows for more flexible and composable operations on tuples. The
CurriedTFold function provides a curried version of TFold, which can be
useful for partial application and functional programming patterns.