test(repository/arangodb): add test for listing features by pubmed ID
Adds a test case `TestListByPublicationId_SuccessPubmed` to verify the
successful retrieval of feature annotations  linked to a specific PubMed
ID using the `ListByPublicationId` method.
The test creates multiple feature annotations, some linked to the target
PubMed ID and some not, then asserts that  only the correctly linked
annotations are returned.