fix(arangodb): preserve original updated at time for tag properties
The `UpdatedAt` field for tag properties was being overwritten with the
current time on every update, regardless of whether the `UpdatedAt`
field in the request was valid. This change ensures that the `UpdatedAt`
field is only updated if a valid `UpdatedAt` is provided in the request,
otherwise it preserves the original `UpdatedAt` time.