feat(arangodb): add convertNewTagToModel function to convert feature.TagPropertyCreate to model.TagPropertyDoc
This function is needed to convert the input struct for creating a new
tag property to the struct that is used for storing tag properties in
the database. It also handles the conversion of the creation timestamp
to the correct format and sets the updated timestamp to the same value.