ci(testcov.yml): update testcov workflow to use coverage.txt and remove GOPROXY
The workflow now uses `coverage.txt` as the coverage file name, which is
the default for the `go test` command. The `GOPROXY` environment
variable has been removed as it is no longer necessary. The `-race` flag
has been added to the `go test` command to enable the race detector,
which helps to find data races in concurrent code.